Add abstract class 'Component' / Complete module5-task1
